requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic-yaml>=1.1.0
PyYAML>=6.0.1
jsonschema
deepdiff

# Azure Services
azure-ai-formrecognizer>=3.3.0
azure-ai-documentintelligence
openai>=1.0.0
azure-identity>=1.15.0

# NLP
spacy
transformers
torch

# LLM Providers
anthropic>=0.8.0

# Document Processing
python-docx>=1.0.0
pdf2image>=1.16.3
Pillow>=10.0.0
shapely

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0

# Logging & Monitoring
structlog>=23.1.0

# Type Checking
types-PyYAML>=6.0.12
types-python-dateutil>=2.8.19

# Optional Dependencies
#redis>=5.0.0  # For caching
#requests>=2.31.0  # For HTTP requests
#python-dateutil>=2.8.2  # For date parsing
